+++ HISTORY.md
@@ -1,3 +1,13 @@
+## 0.5.0
+* Added viewbox()
+* Revised stagger()
+* Allow function values in transition(), transformCss(), transform()
+* Update repo location
+
+## 0.4.0
+* Added Pablo.support
+* Revised transition(), transformCss(), transform()
+
 ## 0.3.8
 * Added transition() for CSS transitions
 * Pass previous element to stagger() callback